requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.13
supabase>=1.0.0
gtts>=2.3.0
textblob>=0.17.1
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
pydub>=0.25.1
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
speechrecognition>=3.10.0
pyaudio>=0.2.11
streamlit-webrtc>=0.47.0
av>=10.0.0
asyncio>=3.4.3
aiofiles>=23.0.0
httpx>=0.24.0
pyttsx3>=2.90
pygame>=2.5.0
pydub>=0.25.1
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
PyPDF2>=3.0.0
python-docx>=0.8.11
speechrecognition>=3.10.0
pyaudio>=0.2.11
streamlit-webrtc>=0.47.0
av>=10.0.0
asyncio>=3.4.3
aiofiles>=23.0.0
httpx>=0.24.0